No more ~/sys/ Signed-off-by: Colin Shea <colin@evaryont.me>
The end-goal was to make a local chroot-able install target (e.g. 'make
install DEST=~/sys/') but a couple things happened: I didn't use it, and
trying to access ~/sys/bin was more of a pain than did it help. Besides,
if I want to install software that's not already packaged, just make a
quickie PKGBUILD (or just run repacman)