Use ${} some more
Vim's syntax highlight was stopping at 'NAME' in '$HOSTNAME' so I fixed
that, also making the whole line a bit more explicit.

The idea: If the syntax highlighter doesn't highlight it properly, there
may be something wrong. If it works anyways, presume that the current
situation is fragile, and try to reinforce what you want/mean to happen,
to hopefully avoid any possibility of mistakes later on.


Former-commit-id: 51a0af1618b45246231668cda029d66c4e261cb9