Revert back to single line prompt
Dunno why, but I lose the last line of output (or the previous prompt)
if I press up, searching through the history. Since not many people have
complained about that, I presume it's related to the method I'm using to
add a newline to the prompt.

Oh well! Just looks like I'm gonna have to embed a newline in the
prompt, literally.