Beginnings of a preparation task
The purpose of the task is to create site-specific directories and files. These
directories and files are used to store various files or settings.

A prime example is the temporary directory for vim (~/.vim/tmp) since it should
exist on every system this configuration resides, but it nor it's contents
should be tracked by git.