Append PATH, update grep, move some configurations
Append PATH if there are any locally installed gems (only for ruby 1.9.1).
Points to the bin/ directory, which is managed by rubygems itself (which then
attempts to load a version of a library and call it's file in bin/).

grep has a new behaviour, which is quite annoying: if called with the recursive
parameter, it will automatically presume you want to search the hard drive, not
standard input (especially when given no hints as to look at files). Odd.

Also, use environment variables to move some more configuration files to the XDG
specification. ~/.inputrc controls readline settings, while I moved mplayer's
"configuration home" so that the defaults, as well as the file-specific
configurations to an XDG dir.


Former-commit-id: 124e65ba6faacc3355bda7ca127ff99bd53c7b07