New function: df, which pretty-fies df
It doesn't do the coloration itself. Rather, it looks for the `dfc` binary and,
if found as well as with no options passed, executes that instead of df itself.
However, if you call it with any parameters, it will execute normal df instead.
No errors or messages are thrown when dfc isn't installed, and will safely fall
back to regular df.


Former-commit-id: 2b1f9866d5302e7424008db81b3e14b80998a004