Switch to xmonad-log-applet
Cleaned up a deprecation warning with ICCCMFocus & takeTopFocus in xmonad>=0.10,
yay less hacks for Java.

The appending of getWellKnownName to xmonad.hs directly is getting to hackish. I
recall some configuration using `xmonad myConfig` even with the dbus stuff, but
it was all contained within myConfig, and there was no additional calls in the
main \-> do block.


Former-commit-id: e0a8c6df023cc84badd2a4ffc9fc2eb906a084a5