Open Build Service Client configuration
I'm going to start using OpenSUSE's Open Build Services to compile & host
packages for my puppet clients, and this is the application to interact with it
from the command line.

Notably, I hope to use it to upload a tarball generated by `makepkg --source` to
OBS and kick off a new build.


Former-commit-id: 1d06cc5662a148969d3cee01dfd3681c0732aa97