Export KDE_SESSION_ID so that python-keyring detects KDE
It's an OpenSUSE oddity, apparently. It's fixed in the latest released version,
but that's not what is available in the repositories.

Also, purple-ize the sudo prompt, having it ask for the target user's name
password, instead of just the simple static "Your", which isn't portable across
different environments. Like, OpenSUSE, which defaults to asking the user for
root's password, not their own.


Former-commit-id: 297df23f163c006668452eafcf08578c3ea22a80