Ignore the private environment zsh.
Stick things like access keys, private data, or other sensitive but
troubling-to-type information in there. Better yet, stick it in an encrypted
wallet and ask KWallet for the information, but not many programs expect data
from an external command, in stead the environment, so here's a good place to
put it.


Former-commit-id: 7ce053e38d54b0f433d4c7edd917b28412f18c88