zsh: psg upgrade, portstat update
portstat now works on machines without netstat already (all modern
distros) instead of me typing the full `sudo ss` shenanigans.

psg now works much better. `--ppid 1` doesn't work as I expected it to.
The pgrep tool does much of what I was attempting anyways, so just rely
on that instead. It's part of the same family of tools as ps anyways,
so if ps is installed pgrep likely is as well.