switch zle to vicmd mode when the prompt is done
This way, no matter if I hit <Enter> in any mode, the
prompt will render as if I were did so in command mode.

This is nice because the character I use for command mode is the filled
in triangle, which provides a nice, clean look for the prompt left
behind.