overhaul how sessions are decided
centralize all of the configuration (autocmds and plugin spec) into a
single file.

create a function to filter the list of buffers, so that I can
selectively ignore some.

also expand the autocmds to be smart about which directories indeed to
create a per-directory session.

giant shout-out to the AstroVim authors for a lot of inspiration on how
to achieve this! this was more than I expected, hah.